engine: extract esm/layers/vo_table.js
